autoencoders_CIFAR100/requirements.txt
Pillow
torch
gradio
torchvision
einops
pytorch_lightning